feat(db) introduce composite_unique DAO constraint
Supports uniqueness check of multiple keys at the generic DAO level.

This is a port of a custom validation that exists for the Plugins
entity in the old DAO, made into a declarative configuration,
that is presented in the schema library as an entity_check.

The Plugins entity in the old DAO currently uses a custom cross-field
validation that essentially performs composite uniqueness constraint: `(name,
api_id, route_id, service_id, consumer_id)`, where any of those ids can be null

note that supporting NULL adds a complication on PostgreSQL:
https://www.dbrnd.com/2016/01/postgresql-composite-unique-constraint-does-not-consider-null-values/

the old DAO implemented this feature in terms of DAO operations,
this new feature replicates that design.